refactor: type data move to one single file
