+++ frontend/src/pages/Login.tsx
@@ -3,7 +3,7 @@ import { useLogin } from "../reducer-context/Login-context";
 import React, { useEffect, useState } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import axios, { AxiosError } from "axios";
-
+import {UserState,Servererror } from "../DataTypes/quiz.types";
 
 import {Button,Box,Grid,TextField,CircularProgress} from "@material-ui/core";
 
@@ -39,16 +39,16 @@ console.log("error",Error);
   }
 
 
-  type User = {
-    _id: string;
-    name: string;
-    email: string;
-  };
-  type Servererror = {
-    errormessage: string;
-  };
+  // type User = {
+  //   _id: string;
+  //   name: string;
+  //   email: string;
+  // };
+  // type Servererror = {
+  //   errormessage: string;
+  // };
 
-  async function verify(): Promise<User | Servererror> {
+  async function verify(): Promise<UserState | Servererror> {
     try {
       let response = await axios.post(
         "https://quiz-backend-demo-1.utpalpati.repl.co/user/infowithcred",
