+++ frontend/src/reducer-context/Reducer-context.tsx
@@ -1,27 +1,27 @@
 import React, { createContext, useContext, useReducer,useEffect,FunctionComponent} from "react";
 import axios from"axios";
-import { quizdata } from "../Data/getQuiz";
-import { Quizdata } from "../Data/quiz.types";
+// import { quizdata } from "../Data/getQuiz";
+import {Contextstate,actiontype,quizstate } from "../DataTypes/quiz.types";
 import { JsxElement } from "typescript";
 
-type statustype = "starting" | "finished" | "Running";
-type UserState = {
-  _id: string;
-  name: string;
-  email: string;
-};
+// type statustype = "starting" | "finished" | "Running";
+// type UserState = {
+//   _id: string;
+//   name: string;
+//   email: string;
+// };
 
-type quizstate = {
-  user: UserState;
-  score: number;
-  status: statustype;
-  currentQsnNo: number;
-  currentquiz: string;
-  correct: number;
-  wrong: number;
-  data: Quizdata;
-  categorydata: string[];
-};
+// type quizstate = {
+//   user: UserState;
+//   score: number;
+//   status: statustype;
+//   currentQsnNo: number;
+//   currentquiz: string;
+//   correct: number;
+//   wrong: number;
+//   data: Quizdata;
+//   categorydata: string[];
+// };
 const initialstate: quizstate = {
   user: { _id: "", name: "", email: "" },
   score: 0,
@@ -31,29 +31,29 @@ const initialstate: quizstate = {
   correct: 0,
   wrong: 0,
   data: {},
-  categorydata: Object.keys(quizdata)
+  categorydata: []
 };
-type actiontype =
-  | {type:  "LOAD"; payload:{data:Quizdata} }
-  | { type: "RESET" }
-  | { type: "INCREMENT_SCORE"; payload: { score: number } }
-  | { type: "DECREMENT_SCORE"; payload: { score: number } }
-  | { type: "SKIP" }
-  | { type: "USER"; payload: UserState }
-  | { type: "CURRENTQUIZ"; payload: string }
-  | {type: "LOGOUT"}
-type Contextstate = {
-  user: UserState;
-  score: number;
-  status: statustype;
-  currentQsnNo: number;
-  currentquiz: string;
-  correct: number;
-  wrong: number;
-  data: Quizdata;
-  categorydata: string[];
-  dispatch: React.Dispatch<actiontype>;
-};
+// type actiontype =
+//   | {type:  "LOAD"; payload:{data:Quizdata} }
+//   | { type: "RESET" }
+//   | { type: "INCREMENT_SCORE"; payload: { score: number } }
+//   | { type: "DECREMENT_SCORE"; payload: { score: number } }
+//   | { type: "SKIP" }
+//   | { type: "USER"; payload: UserState }
+//   | { type: "CURRENTQUIZ"; payload: string }
+//   | {type: "LOGOUT"}
+// type Contextstate = {
+//   user: UserState;
+//   score: number;
+//   status: statustype;
+//   currentQsnNo: number;
+//   currentquiz: string;
+//   correct: number;
+//   wrong: number;
+//   data: Quizdata;
+//   categorydata: string[];
+//   dispatch: React.Dispatch<actiontype>;
+// };
 
 export const Reducercontext = createContext({} as Contextstate);
 
@@ -86,8 +86,7 @@ function quizreducer(state: quizstate, action: actiontype): quizstate {
         currentquiz: "",
         correct: 0,
         wrong: 0,
-        data: quizdata,
-        categorydata: Object.keys(quizdata)
+        categorydata: Object.keys(state.data)
       }
     case "INCREMENT_SCORE":
       if (
