+++ frontend/src/utils/score-updates.ts
@@ -1,4 +1,4 @@
-import { question, option } from "../Data/quiz.types";
+import { question, option } from "../DataTypes/quiz.types";
 
 function update(score: number, qsn: question, opt: option) {
   return opt.isRight ? score + qsn.points : score;
